refactor(go-template): extract memo initial-value computation into memo/memo-compute.ts
Move the six mutually-recursive memo-value functions out of the adapter
class into pure free functions (Roadmap B): computeMemoInitialValue,
computeMemoInitialValueOrNull, memoInitialFromParsedBody,
computeComparisonTernaryGo, resolveComparisonOperandGo,
resolveGetterValueAsGo. They read state.currentMemos /
state.moduleStringConsts and delegate to the already-extracted value /
type / template-interp / memo-value modules.

GoEmitContext gains extractPropFallback (parallel to the existing
extractPropNameFromInitialValue), the one adapter-resident parser the
core calls back into. The now-unused EMPTY_PROP_FALLBACK_VARS static is
removed (all callers moved into modules).

Adapter is now ~6.1k lines (from ~7.6k at the start of Roadmap B).
Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
